+++ sso/views.py
@@ -1,6 +1,7 @@
 import requests
 from rest_framework import generics
 from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated
 
 from django.conf import settings
 
@@ -78,5 +79,18 @@ class SSOBusinessVerifyCodeView(generics.GenericAPIView):
             email=serializer.validated_data['email'],
             form_url=self.request.path
         )
-
         return helpers.response_factory(upstream_response=upstream_response)
+
+
+class LessonCompletedAPIView(generics.GenericAPIView):
+    permission_classes = [IsAuthenticated]
+
+    def post(self, request, *args, **kwargs):
+        session_id = request.user.session_id
+        response = helpers.set_lesson_completed(session_id, kwargs['lesson'])
+        return Response(status=200, data=response)
+
+    def get(self, request, *args, **kwargs):
+        session_id = request.user.session_id
+        response = helpers.get_lesson_completed(session_id, kwargs['lesson'])
+        return Response(status=200, data=response)