+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 - AA-129 - Beta environment for GREAT.gov
 - AA-129 - wagtail import export for Beta environment
+- GP2-36 - lesson completed functionality
 
 ## Pre-release